@@ -1,47 +1,52 @@
+cargo-features = ["profile-rustflags", "trim-paths"]
+
 [package]
 name = "cursor-api"
-version = "0.1.3-rc.4.3"
-edition = "2021"
+version = "0.1.3-rc.5.2.5"
+edition = "2024"
 authors = ["wisdgod <nav@wisdgod.com>"]
 description = "OpenAI format compatibility layer for the Cursor API"
 repository = "https://github.com/wisdgod/cursor-api"
 
 [build-dependencies]
-prost-build = "0.13.4"
-sha2 = { version = "0.10.8", default-features = false }
-serde_json = "1.0.134"
+prost-build = "^0.13"
+sha2 = { version = "^0.10.8", default-features = false }
+serde_json = "^1.0"
 
 [dependencies]
-axum = { version = "0.8.1", features = ["json"] }
-base64 = { version = "0.22.1", default-features = false, features = ["std"] }
-# brotli = { version = "7.0.0", default-features = false, features = ["std"] }
-bytes = "1.9.0"
-chrono = { version = "0.4.39", default-features = false, features = ["std", "clock", "now", "serde", "rkyv-64"] }
-dotenvy = "0.15.7"
-flate2 = { version = "1.0.35", default-features = false, features = ["rust_backend"] }
-futures = { version = "0.3.31", default-features = false, features = ["std"] }
-gif = { version = "0.13.1", default-features = false, features = ["std"] }
-hex = { version = "0.4.3", default-features = false, features = ["std"] }
-image = { version = "0.25.5", default-features = false, features = ["jpeg", "png", "gif", "webp"] }
-memmap2 = "0.9.5"
-# openssl = { version = "0.10.68", features = ["vendored"] }
-parking_lot = "0.12.3"
-paste = "1.0.15"
-prost = "0.13.4"
-rand = { version = "0.8.5", default-features = false, features = ["std", "std_rng"] }
-regex = { version = "1.11.1", default-features = false, features = ["std", "perf"] }
-reqwest = { version = "0.12.12", default-features = false, features = ["gzip", "brotli", "json", "stream", "socks", "__tls", "charset", "default-tls", "h2", "http2", "macos-system-configuration"] }
-rkyv = { version = "0.7.45", default-features = false, features = ["alloc", "std", "bytecheck", "size_64", "validation", "std"] }
-serde = { version = "1.0.217", default-features = false, features = ["std", "derive"] }
-serde_json = { package = "sonic-rs", version = "0.3.17" }
-# serde_json = "1.0.137"
-sha2 = { version = "0.10.8", default-features = false }
-sysinfo = { version = "0.33.1", default-features = false, features = ["system"] }
-tokio = { version = "1.43.0", features = ["rt-multi-thread", "macros", "net", "sync", "time", "fs", "signal"] }
-tokio-stream = { version = "0.1.17", features = ["time"] }
-tower-http = { version = "0.6.2", features = ["cors", "limit"] }
-url = { version = "2.5.4", default-features = false }
-uuid = { version = "1.12.1", features = ["v4"] }
+axum = { version = "^0.8", default-features = false, features = ["http1", "http2", "json", "tokio", "query"] }
+base64 = { version = "^0.22", default-features = false, features = ["std"] }
+# brotli = { version = "^7.0", default-features = false, features = ["std"] }
+bytes = "^1.10"
+chrono = { version = "^0.4", default-features = false, features = ["std", "clock", "now", "serde", "rkyv-64"] }
+chrono-tz = { version = "^0.10", features = ["serde"] }
+dotenvy = "^0.15"
+flate2 = { version = "1", default-features = false, features = ["rust_backend"] }
+futures = { version = "^0.3", default-features = false, features = ["std"] }
+gif = { version = "^0.13", default-features = false, features = ["std"] }
+hex = { version = "^0.4", default-features = false, features = ["std"] }
+http = "1"
+image = { version = "^0.25", default-features = false, features = ["jpeg", "png", "gif", "webp"] }
+lasso = { version = "^0.7", features = ["inline-more", "multi-threaded"] }
+memmap2 = "^0.9"
+# openssl = { version = "^0.10", features = ["vendored"] }
+parking_lot = "^0.12"
+paste = "^1.0"
+prost = "^0.13"
+prost-types = "^0.13"
+rand = { version = "^0.9", default-features = false, features = ["thread_rng"] }
+reqwest = { version = "^0.12", default-features = false, features = ["gzip", "brotli", "json", "stream", "socks", "__tls", "charset", "rustls-tls-webpki-roots", "macos-system-configuration"] }
+rkyv = { version = "^0.7", default-features = false, features = ["alloc", "std", "bytecheck", "size_64", "validation", "std"] }
+serde = { version = "^1.0", default-features = false, features = ["std", "derive", "rc"] }
+serde_json = { package = "sonic-rs", version = "0.5" }
+# serde_json = "^1.0"
+sha2 = { version = "^0.10", default-features = false }
+sysinfo = { version = "^0.34", default-features = false, features = ["system"] }
+tokio = { version = "^1.43", features = ["rt-multi-thread", "macros", "net", "sync", "time", "fs", "signal"] }
+# tokio-stream = { version = "^0.1", features = ["time"] }
+tower-http = { version = "^0.6", features = ["cors", "limit"] }
+url = { version = "^2.5", default-features = false, features = ["serde"] }
+uuid = { version = "^1.14", features = ["v4"] }
 
 [profile.release]
 lto = true
