@@ -1,3 +1,5 @@
+use std::borrow::Cow;
+
 use super::ErrorResponse;
 
 pub enum ChatError {
@@ -13,22 +15,22 @@ impl ChatError {
         let (error, message) = match self {
             ChatError::ModelNotSupported(model) => (
                 "model_not_supported",
-                format!("Model '{}' is not supported", model),
+                format!("Model '{model}' is not supported"),
             ),
             ChatError::EmptyMessages => (
                 "empty_messages",
                 "Message array cannot be empty".to_string(),
             ),
             ChatError::NoTokens => ("no_tokens", "No available tokens".to_string()),
-            ChatError::RequestFailed(err) => ("request_failed", format!("Request failed: {}", err)),
+            ChatError::RequestFailed(err) => ("request_failed", format!("Request failed: {err}")),
             ChatError::Unauthorized => ("unauthorized", "Invalid authorization token".to_string()),
         };
 
         ErrorResponse {
-          status: super::ApiStatus::Error,
-          code: None,
-          error: Some(error.to_string()),
-          message: Some(message),
+            status: super::ApiStatus::Error,
+            code: None,
+            error: Some(Cow::Borrowed(error)),
+            message: Some(Cow::Owned(message)),
         }
     }
 }
