@@ -1,14 +1,17 @@
-use base64::{engine::general_purpose::STANDARD as BASE64, Engine as _};
-use rand::Rng;
+use base64::{Engine as _, engine::general_purpose::URL_SAFE_NO_PAD as BASE64};
 use sha2::{Digest, Sha256};
 
+#[inline]
 pub fn generate_hash() -> String {
-    let random_bytes = rand::thread_rng().gen::<[u8; 32]>();
-    let mut hasher = Sha256::new();
-    hasher.update(random_bytes);
-    hex::encode(hasher.finalize())
+    use rand::Rng as _;
+    let mut v = rand::rng().random::<[u8; 32]>();
+    if *crate::app::lazy::SAFE_HASH {
+        v = Sha256::new().chain_update(v).finalize().into();
+    }
+    hex::encode(v)
 }
 
+#[inline]
 fn obfuscate_bytes(bytes: &mut [u8]) {
     let mut prev: u8 = 165;
     for (idx, byte) in bytes.iter_mut().enumerate() {
@@ -18,6 +21,7 @@ fn obfuscate_bytes(bytes: &mut [u8]) {
     }
 }
 
+#[inline]
 fn deobfuscate_bytes(bytes: &mut [u8]) {
     let mut prev: u8 = 165;
     for (idx, byte) in bytes.iter_mut().enumerate() {
@@ -27,14 +31,19 @@ fn deobfuscate_bytes(bytes: &mut [u8]) {
     }
 }
 
-pub fn generate_timestamp_header() -> String {
-    let timestamp = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap()
-        .as_secs()
-        / 1_000;
+pub fn generate_timestamp_header() -> [u8; 8] {
+    static CACHE: std::sync::LazyLock<parking_lot::Mutex<(u64, [u8; 8])>> =
+        std::sync::LazyLock::new(|| parking_lot::Mutex::new((0, [0u8; 8])));
+    let timestamp = super::now_secs() / 1_000;
 
-    let mut timestamp_bytes = vec![
+    let mut guard = CACHE.lock();
+    if guard.0 == timestamp {
+        return guard.1;
+    } else {
+        guard.0 = timestamp;
+    }
+
+    let mut timestamp_bytes = [
         ((timestamp >> 8) & 0xFF) as u8,
         (0xFF & timestamp) as u8,
         ((timestamp >> 24) & 0xFF) as u8,
@@ -44,17 +53,23 @@ pub fn generate_timestamp_header() -> String {
     ];
 
     obfuscate_bytes(&mut timestamp_bytes);
-    BASE64.encode(&timestamp_bytes)
+    let mut result = [0u8; 8];
+    let _ = BASE64.encode_slice(&timestamp_bytes, &mut result);
+    guard.1 = result;
+    result
 }
 
+#[inline]
 pub fn generate_checksum(device_id: &str, mac_addr: Option<&str>) -> String {
-    let encoded = generate_timestamp_header();
+    let timestamp_header = generate_timestamp_header();
+    let encoded = unsafe { str::from_utf8_unchecked(&timestamp_header) };
     match mac_addr {
-        Some(mac) => format!("{}{}/{}", encoded, device_id, mac),
-        None => format!("{}{}", encoded, device_id),
+        Some(mac) => format!("{encoded}{device_id}/{mac}"),
+        None => format!("{encoded}{device_id}"),
     }
 }
 
+#[inline]
 pub fn generate_checksum_with_default() -> String {
     generate_checksum(&generate_hash(), Some(&generate_hash()))
 }
@@ -72,7 +87,7 @@ pub fn generate_checksum_with_repair(checksum: &str) -> String {
     for (i, &b) in bytes.iter().enumerate() {
         let valid = match (len, i) {
             // 通用字符校验（排除非法字符）
-            (_, _) if !b.is_ascii_alphanumeric() && b != b'/' && b != b'+' && b != b'=' => false,
+            (_, _) if !b.is_ascii_alphanumeric() && b != b'/' && b != b'-' && b != b'_' => false,
 
             // 72字节格式：时间戳(8) + 设备哈希(64)
             (72, 8..=71) => b.is_ascii_hexdigit(),
@@ -98,25 +113,25 @@ pub fn generate_checksum_with_repair(checksum: &str) -> String {
         }
     }
 
+    let timestamp_header = generate_timestamp_header();
+    let encoded = unsafe { str::from_utf8_unchecked(&timestamp_header) };
+
     // 校验通过后构造结果
     match len {
         72 => format!(
-            "{}{}/{}",
-            generate_timestamp_header(),
-            unsafe { std::str::from_utf8_unchecked(&bytes[8..]) },
+            "{encoded}{}/{}",
+            unsafe { std::str::from_utf8_unchecked(bytes.get_unchecked(8..)) },
             generate_hash()
         ),
         129 => format!(
-            "{}{}/{}",
-            generate_timestamp_header(),
-            unsafe { std::str::from_utf8_unchecked(&bytes[..64]) },
-            unsafe { std::str::from_utf8_unchecked(&bytes[65..]) }
+            "{encoded}{}/{}",
+            unsafe { std::str::from_utf8_unchecked(bytes.get_unchecked(..64)) },
+            unsafe { std::str::from_utf8_unchecked(bytes.get_unchecked(65..)) }
         ),
         137 => format!(
-            "{}{}/{}",
-            generate_timestamp_header(),
-            unsafe { std::str::from_utf8_unchecked(&bytes[8..72]) },
-            unsafe { std::str::from_utf8_unchecked(&bytes[73..]) }
+            "{encoded}{}/{}",
+            unsafe { std::str::from_utf8_unchecked(bytes.get_unchecked(8..72)) },
+            unsafe { std::str::from_utf8_unchecked(bytes.get_unchecked(73..)) }
         ),
         _ => unreachable!(),
     }
@@ -131,17 +146,21 @@ pub fn extract_time_ks(timestamp_base64: &str) -> Option<u64> {
 
     deobfuscate_bytes(&mut timestamp_bytes);
 
-    if timestamp_bytes[0] != timestamp_bytes[4] || timestamp_bytes[1] != timestamp_bytes[5] {
-        return None;
-    }
+    unsafe {
+        if timestamp_bytes.get_unchecked(0) != timestamp_bytes.get_unchecked(4)
+            || timestamp_bytes.get_unchecked(1) != timestamp_bytes.get_unchecked(5)
+        {
+            return None;
+        }
 
-    // 使用后四位还原 timestamp
-    Some(
-        ((timestamp_bytes[2] as u64) << 24)
-            | ((timestamp_bytes[3] as u64) << 16)
-            | ((timestamp_bytes[4] as u64) << 8)
-            | (timestamp_bytes[5] as u64),
-    )
+        // 使用后四位还原 timestamp
+        Some(
+            ((*timestamp_bytes.get_unchecked(2) as u64) << 24)
+                | ((*timestamp_bytes.get_unchecked(3) as u64) << 16)
+                | ((*timestamp_bytes.get_unchecked(4) as u64) << 8)
+                | (*timestamp_bytes.get_unchecked(5) as u64),
+        )
+    }
 }
 
 pub fn validate_checksum(checksum: &str) -> bool {
@@ -157,7 +176,7 @@ pub fn validate_checksum(checksum: &str) -> bool {
     for (i, &b) in bytes.iter().enumerate() {
         let valid = match (len, i) {
             // 通用字符校验（排除非法字符）
-            (_, _) if !b.is_ascii_alphanumeric() && b != b'/' && b != b'+' && b != b'=' => false,
+            (_, _) if !b.is_ascii_alphanumeric() && b != b'/' && b != b'-' && b != b'_' => false,
 
             // 格式校验
             (72, 0..=7) => true, // 时间戳部分（由extract_time_ks验证）
@@ -177,16 +196,9 @@ pub fn validate_checksum(checksum: &str) -> bool {
     }
 
     // 统一时间戳验证（无需分层）
-    let time_valid = extract_time_ks(&checksum[..8]).is_some();
+    let time_valid = extract_time_ks(unsafe { checksum.get_unchecked(..8) }).is_some();
 
-    // 附加MAC哈希长度校验（仅137字符需要）
-    let mac_hash_valid = if len == 137 {
-        checksum[73..].len() == 64 // 确保MAC哈希长度为64
-    } else {
-        true // 72字符无需此检查
-    };
-
-    time_valid && mac_hash_valid
+    time_valid
 }
 
 /// 从校验通过的checksum中提取哈希值（需先通过validate_checksum验证）
@@ -201,14 +213,14 @@ pub fn extract_hashes(checksum: &str) -> Option<(Vec<u8>, Vec<u8>)> {
     match checksum.len() {
         72 => {
             // 格式：8字节时间戳 + 64字节设备哈希
-            let device_hash = hex::decode(&checksum[8..]).ok()?; // 8..72
+            let device_hash = hex::decode(unsafe { checksum.get_unchecked(8..) }).ok()?; // 8..72
             Some((device_hash, Vec::new()))
         }
         137 => {
             // 格式：8时间戳 + 64设备哈希 + '/' + 64MAC哈希
             // 直接按固定位置切割（validate_checksum已确保索引72是'/'）
-            let device_hash = hex::decode(&checksum[8..72]).ok()?;
-            let mac_hash = hex::decode(&checksum[73..]).ok()?; // 73..137
+            let device_hash = hex::decode(unsafe { checksum.get_unchecked(8..72) }).ok()?;
+            let mac_hash = hex::decode(unsafe { checksum.get_unchecked(73..) }).ok()?; // 73..137
             Some((device_hash, mac_hash))
         }
         // validate_checksum已过滤其他长度，此处应为不可达代码
