@@ -1,3 +1,3 @@
+pub mod client;
 pub mod model;
 pub mod utils;
-pub mod client;
