@@ -1,15 +1,15 @@
 use chrono::{DateTime, Local};
-use serde::{Deserialize, Serialize};
 use rkyv::{Archive, Deserialize as RkyvDeserialize, Serialize as RkyvSerialize};
+use serde::{Deserialize, Serialize};
 
 #[derive(Serialize)]
 #[serde(untagged)]
 pub enum GetUserInfo {
-    Usage(TokenProfile),
+    Usage(Box<TokenProfile>),
     Error { error: String },
 }
 
-#[derive(Serialize, Clone, Archive, RkyvDeserialize, RkyvSerialize)]
+#[derive(Deserialize, Serialize, Clone, Archive, RkyvDeserialize, RkyvSerialize)]
 pub struct TokenProfile {
     pub usage: UsageProfile,
     pub user: UserProfile,
@@ -28,52 +28,62 @@ pub enum MembershipType {
     Enterprise,
 }
 
+impl std::str::FromStr for MembershipType {
+    type Err = ();
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "free" => Ok(MembershipType::Free),
+            "free_trial" => Ok(MembershipType::FreeTrial),
+            "pro" => Ok(MembershipType::Pro),
+            "enterprise" => Ok(MembershipType::Enterprise),
+            _ => Err(()),
+        }
+    }
+}
+
 #[derive(Deserialize, Serialize, Clone, Archive, RkyvDeserialize, RkyvSerialize)]
 pub struct StripeProfile {
-    #[serde(rename(deserialize = "membershipType"))]
+    #[serde(alias = "membershipType")]
     pub membership_type: MembershipType,
-    #[serde(
-        rename(deserialize = "paymentId"),
-        default,
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(alias = "paymentId", default, skip_serializing_if = "Option::is_none")]
     pub payment_id: Option<String>,
-    #[serde(rename(deserialize = "daysRemainingOnTrial"))]
+    #[serde(alias = "daysRemainingOnTrial")]
     pub days_remaining_on_trial: u32,
 }
 
 #[derive(Deserialize, Serialize, Clone, Archive, RkyvDeserialize, RkyvSerialize)]
 pub struct ModelUsage {
-    #[serde(rename(deserialize = "numRequests", serialize = "requests"))]
+    #[serde(
+        alias = "numRequests",
+        alias = "requests",
+        rename(serialize = "requests")
+    )]
     pub num_requests: u32,
     #[serde(
-        rename(deserialize = "numRequestsTotal"),
+        alias = "numRequestsTotal",
         default,
         skip_serializing_if = "Option::is_none"
     )]
-    pub requests_total: Option<u32>,
-    #[serde(rename(deserialize = "numTokens", serialize = "tokens"))]
+    pub total_requests: Option<u32>,
+    #[serde(alias = "numTokens", alias = "tokens", rename(serialize = "tokens"))]
     pub num_tokens: u32,
-    #[serde(
-        rename(deserialize = "maxRequestUsage"),
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(alias = "maxRequestUsage", skip_serializing_if = "Option::is_none")]
     pub max_requests: Option<u32>,
-    #[serde(
-        rename(deserialize = "maxTokenUsage"),
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(alias = "maxTokenUsage", skip_serializing_if = "Option::is_none")]
     pub max_tokens: Option<u32>,
 }
 
 #[derive(Deserialize, Serialize, Clone, Archive, RkyvDeserialize, RkyvSerialize)]
 pub struct UsageProfile {
-    #[serde(rename(deserialize = "gpt-4"))]
+    #[serde(alias = "gpt-4")]
     pub premium: ModelUsage,
-    #[serde(rename(deserialize = "gpt-3.5-turbo"))]
+    #[serde(alias = "gpt-3.5-turbo")]
     pub standard: ModelUsage,
-    #[serde(rename(deserialize = "gpt-4-32k"))]
+    #[serde(alias = "gpt-4-32k")]
     pub unknown: ModelUsage,
+    #[serde(alias = "startOfMonth")]
+    pub start_of_month: DateTime<Local>,
 }
 
 #[derive(Deserialize, Serialize, Clone, Archive, RkyvDeserialize, RkyvSerialize)]
@@ -81,7 +91,7 @@ pub struct UserProfile {
     pub email: String,
     // pub email_verified: bool,
     pub name: String,
-    #[serde(rename(serialize = "id"))]
+    #[serde(alias = "id", rename(serialize = "id"))]
     pub sub: String,
     pub updated_at: DateTime<Local>,
     // Image link, rendered in /logs?
