@@ -1,23 +1,6 @@
-use super::generate_checksum_with_repair;
-use crate::app::{
-    constant::{COMMA, EMPTY_STRING},
-    lazy::TOKEN_LIST_FILE,
-    model::TokenInfo,
-};
 use crate::common::model::token::TokenPayload;
-use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine};
-use chrono::{DateTime, Local, TimeZone};
-
-// 规范化文件内容并写入
-fn normalize_and_write(content: &str, file_path: &str) -> String {
-    let normalized = content.replace("\r\n", "\n");
-    if normalized != content {
-        if let Err(e) = std::fs::write(file_path, &normalized) {
-            eprintln!("警告: 无法更新规范化的文件: {}", e);
-        }
-    }
-    normalized
-}
+use base64::{Engine, engine::general_purpose::URL_SAFE_NO_PAD};
+use chrono::{DateTime, Local, TimeZone as _};
 
 // 解析token
 pub fn parse_token(token_part: &str) -> String {
@@ -38,81 +21,40 @@ pub fn parse_token(token_part: &str) -> String {
     }
 }
 
-// Token 加载函数
-pub fn load_tokens() -> Vec<TokenInfo> {
-    let token_list_file = TOKEN_LIST_FILE.as_str();
+// Token 加载函数，支持从字符串内容加载
+// pub fn load_tokens_from_content(content: &str) -> Vec<TokenInfo> {
+//     let token_map: std::collections::HashMap<String, String> = content
+//         .lines()
+//         .filter_map(|line| {
+//             let line = line.trim();
+//             if line.is_empty() || line.starts_with('#') {
+//                 return None;
+//             }
 
-    // 确保文件存在
-    if !std::path::Path::new(&token_list_file).exists() {
-        if let Err(e) = std::fs::write(&token_list_file, EMPTY_STRING) {
-            eprintln!("警告: 无法创建文件 '{}': {}", &token_list_file, e);
-        }
-    }
+//             let parts: Vec<&str> = line.split(COMMA).collect();
+//             match parts[..] {
+//                 [token_part, checksum] => {
+//                     let token = parse_token(token_part);
+//                     Some((token, generate_checksum_with_repair(checksum)))
+//                 }
+//                 _ => {
+//                     eprintln!("警告: 忽略无效的token-list行: {}", line);
+//                     None
+//                 }
+//             }
+//         })
+//         .collect();
 
-    // 读取和规范化 token-list 文件
-    let token_map: std::collections::HashMap<String, String> =
-        match std::fs::read_to_string(&token_list_file) {
-            Ok(content) => {
-                let normalized = normalize_and_write(&content, &token_list_file);
-                normalized
-                    .lines()
-                    .filter_map(|line| {
-                        let line = line.trim();
-                        if line.is_empty() || line.starts_with('#') {
-                            return None;
-                        }
-
-                        let parts: Vec<&str> = line.split(COMMA).collect();
-                        match parts[..] {
-                            [token_part, checksum] => {
-                                let token = parse_token(token_part);
-                                Some((token, generate_checksum_with_repair(checksum)))
-                            }
-                            _ => {
-                                eprintln!("警告: 忽略无效的token-list行: {}", line);
-                                None
-                            }
-                        }
-                    })
-                    .collect()
-            }
-            Err(e) => {
-                eprintln!("警告: 无法读取token-list文件: {}", e);
-                std::collections::HashMap::new()
-            }
-        };
-
-    // 更新 token-list 文件
-    let token_list_content = token_map
-        .iter()
-        .map(|(token, checksum)| format!("{},{}", token, checksum))
-        .collect::<Vec<_>>()
-        .join("\n");
-
-    if let Err(e) = std::fs::write(&token_list_file, token_list_content) {
-        eprintln!("警告: 无法更新token-list文件: {}", e);
-    }
-
-    // 转换为 TokenInfo vector
-    token_map
-        .into_iter()
-        .map(|(token, checksum)| TokenInfo {
-            token: token.clone(),
-            checksum,
-            profile: None,
-        })
-        .collect()
-}
-
-pub fn write_tokens(token_infos: &[TokenInfo], file_path: &str) -> std::io::Result<()> {
-    let content = token_infos
-        .iter()
-        .map(|info| format!("{},{}", info.token, info.checksum))
-        .collect::<Vec<String>>()
-        .join("\n");
-
-    std::fs::write(file_path, content)
-}
+//     token_map
+//         .into_iter()
+//         .map(|(token, checksum)| TokenInfo {
+//             token,
+//             checksum,
+//             profile: None,
+//             tags: None,
+//         })
+//         .collect()
+// }
 
 pub(super) const HEADER_B64: &str = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9";
 pub(super) const ISSUER: &str = "https://authentication.cursor.sh";
@@ -181,10 +123,10 @@ pub fn validate_token(token: &str) -> bool {
     }
 
     // 验证过期时间
-    let current_time = chrono::Utc::now().timestamp();
-    if current_time > payload.exp {
-        return false;
-    }
+    // let current_time = chrono::Utc::now().timestamp();
+    // if current_time > payload.exp {
+    //     return false;
+    // }
 
     // 验证发行者
     if payload.iss != ISSUER {
@@ -241,8 +183,14 @@ pub fn extract_user_id(token: &str) -> Option<String> {
     )
 }
 
+#[derive(serde::Serialize)]
+pub struct JwtTime {
+    pub iat: DateTime<Local>,
+    pub exp: DateTime<Local>,
+}
+
 // 从 JWT token 中提取 time 字段
-pub fn extract_time(token: &str) -> Option<DateTime<Local>> {
+pub fn extract_time(token: &str) -> Option<JwtTime> {
     // JWT token 由3部分组成，用 . 分隔
     let parts: Vec<&str> = token.split('.').collect();
     if parts.len() != 3 {
@@ -254,6 +202,8 @@ pub fn extract_time(token: &str) -> Option<DateTime<Local>> {
         Ok(decoded) => decoded,
         Err(_) => return None,
     };
+
+    drop(parts);
 
     // 将 payload 转换为字符串
     let payload_str = match String::from_utf8(payload) {
@@ -267,10 +217,12 @@ pub fn extract_time(token: &str) -> Option<DateTime<Local>> {
         Err(_) => return None,
     };
 
-    // 提取时间戳并转换为本地时间
-    payload
+    let iat = payload
         .time
         .parse::<i64>()
         .ok()
-        .and_then(|timestamp| Local.timestamp_opt(timestamp, 0).single())
+        .and_then(|ts| Local.timestamp_opt(ts, 0).single())?;
+    let exp = Local.timestamp_opt(payload.exp, 0).single()?;
+
+    Some(JwtTime { iat, exp })
 }
